requirements.txt
tensorflow
numpy
pillow
scikit-learn
matplotlib==3.7.1
seaborn
pandas
pycocotools